+++ src/index.php
@@ -1,0 +1,8 @@
+<?php
+
+require_once 'vendor/autoload.php';
+
+use Humble23\CartolaFcClient\CartolaClient;
+
+$client = new CartolaClient();
+