+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+All notable changes to `cartola-fc-api` will be documented in this file.
 
-## 1.0.0 - 202X-XX-XX
+## 1.0.0 - 2021-12-30
 
 - initial release
+- Added Guzzle Http Client